create separate tests project
