+++ CodeReviews.Console.CodingTracker.aneevel/CodeReviews.Console.CodingTracker.aneevel.slnx
@@ -1,3 +1,4 @@
 <Solution>
+  <Project Path="../Tests/Tests.csproj" Id="d05e7d66-2c85-4055-ac8c-e5641941ac18" />
   <Project Path="CodeReviews.Console.CodingTracker.aneevel.csproj" />
 </Solution>
