+++ CodeReviews.Console.CodingTracker.aneevel/Services/CodingSessionService.cs
@@ -3,13 +3,13 @@ using CodeReviews.Console.CodingTracker.aneevel.Models;
 
 namespace CodeReviews.Console.CodingTracker.aneevel.Services;
 
-internal class CodingSessionService(ICodingSessionRepository repository) : ICodingSessionService
+public class CodingSessionService(ICodingSessionRepository repository) : ICodingSessionService
 {
     public List<CodingSession> GetCodingSessions()
     {
         return repository.GetCodingSessions();
     }
-    
+
     public int InsertCodingSession(CodingSession codingSession)
     {
         return repository.InsertCodingSession(codingSession);
@@ -17,7 +17,7 @@ internal class CodingSessionService(ICodingSessionRepository repository) : ICodi
 
     public int UpdateCodingSession(int id, DateTime startTime, DateTime endTime, TimeSpan duration)
     {
-       return repository.UpdateCodingSession(id, startTime, endTime, duration); 
+        return repository.UpdateCodingSession(id, startTime, endTime, duration);
     }
 
     public int DeleteCodingSession(int id)
